Update kento_radio.sp
This should fix the bug that emit the "death" sound for the whole team. Also fix the bug that the groan (of death)'s position to stay at the corpse's origin, not following the receiver's origin.

`clients` array of SoundHook "Event_SoundPlayed" already has the clients supposed to receive the radio (teammate only) or the groan (everyone in the room), so using the array directly is okay, I think?